CP-54351 move SRCommand from /opt
