+++ drivers/ISOSR.py
@@ -19,7 +19,7 @@
 
 import SR
 import VDI
-import SRCommand
+from sm import SRCommand
 from sm.core import util
 from sm import nfs
 import os
