+++ drivers/RawISCSISR.py
@@ -19,7 +19,7 @@
 #
 
 import SR
-import SRCommand
+from sm import SRCommand
 import BaseISCSI
 from sm import LUNperVDI
 from sm.core import util
