+++ drivers/SHMSR.py
@@ -17,7 +17,7 @@
 
 import SR
 import VDI
-import SRCommand
+from sm import SRCommand
 from sm.core import util
 import os
 from sm.core import xs_errors
